requirements.txt
amqp==2.3.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
djangorestframework==3.9.0
drf-yasg==1.11.1
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2
MarkupSafe==1.1.0
psycopg2-binary==2.7.6.1
pytz==2018.7
redis==3.0.1
requests==2.21.0
ruamel.yaml==0.15.81
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
